% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{age_wtd_ifr}
\alias{age_wtd_ifr}
\title{Age-weighted IFR of COVID-19}
\usage{
age_wtd_ifr(locations)
}
\arguments{
\item{locations}{list of country names. Note that no checks have
been implemented to test that the country names are valid.
Use names as in ECDC COVID-19 data, so United_Kingdom instead of
any other variation.}
}
\value{
a named list with samples from age-weighted IFR for each
location
}
\description{
COVID-19 IFR adjusted for age compoisition of a country
}
\author{
Sangeeta Bhatia
}
