% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_epidemic_phase2}
\alias{assign_epidemic_phase2}
\title{Assign epidemic phase}
\usage{
assign_epidemic_phase2(rt)
}
\arguments{
\item{rt}{a vector of samples from the
posterior distribution}
}
\value{
phase
}
\description{
Use Epinow definitions
}
\author{
Sangeeta Bhatia
}
