% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_predictions}
\alias{cap_predictions}
\title{Cap predictions to twice the observed for visualisation}
\usage{
cap_predictions(pred)
}
\value{
data.frame with capped predictions
}
\description{
Cap predictions to twice the observed for visualisation
}
\author{
Sangeeta Bhatia
}
