% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_continent_mapping.R
\docType{data}
\name{country_continent_mapping}
\alias{country_continent_mapping}
\title{Country to continent mapping}
\format{
A data frame with 210 rows and 2 variables:
\describe{
  \item{Countries and territories}{name of country or territory}
  \item{continent}{continent}
}
}
\usage{
country_continent_mapping
}
\description{
A dataset containing a list of countries and corresponding
continents. We use this because it lists countries in North America
and South America separately.
}
\keyword{datasets}
