% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{customise_for_rows}
\alias{customise_for_rows}
\title{Customise plots to be aligned in rows in the manuscript}
\usage{
customise_for_rows(plots, in_rows)
}
\arguments{
\item{plots}{list of ggplot2 objects}

\item{in_rows}{list of indices of plots that will be put in rows}
}
\description{
Customise plots to be aligned in rows in the manuscript
}
\details{
The very first plot will have legend but no x-axis text;
all plots in the middle will
have no legend and no x-axis text, and the very last plot will have
no legend but x-axis text
}
