% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlaps}
\alias{overlaps}
\title{Check if two intervals overlap}
\usage{
overlaps(x1, x2, digits)
}
\arguments{
\item{x1}{a numeric vector of length 2}

\item{x2}{a numeric vector of length 2}

\item{digits}{integer indicating the number of decimal places
to be used}
}
\value{
TRUE if x1 and x2 overlap
}
\description{
Check if two intervals overlap
}
\author{
Sangeeta Bhatia
}
