% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{restimates_linegraph}
\alias{restimates_linegraph}
\title{Rt quantile plot}
\usage{
restimates_linegraph(
  df,
  group_var,
  date_breaks = "1 month",
  date_labels = "\%d - \%b"
)
}
\arguments{
\item{df}{data.frame of quantiles for Rt. Columns should be called
dates, `2.5%`, `97.5%` and `50%`}

\item{group_var}{variable to group by; forecast_date or
forecast_week}
}
\value{
ggplot2 object
}
\description{
Rt quantile plot
}
\author{
Sangeeta Bhatia
}
