% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scale_date_manuscript}
\alias{scale_date_manuscript}
\title{Defines a date scale to be used for all figures.}
\usage{
scale_date_manuscript(date_breaks, date_labels, xmin)
}
\arguments{
\item{date_breaks}{breaks for x-axis, Defaults to 1 month}

\item{date_labels}{labels, default is day-month}
}
\description{
Defines a date scale to be used for all figures.
}
