% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_to_incid}
\alias{ts_to_incid}
\title{Convert time-series to incidence object}
\usage{
ts_to_incid(ts, date_col, case_col)
}
\value{

}
\description{
incidence package accepts a linelist like object (list of dates)
and converts them into an incid object, which is needed by
projections package. To use projections package with a time-series
we convert time-series object into incid object.
}
\author{
Sangeeta Bhatia
}
